+++ jellyfin-core/build.gradle.kts
@@ -88,13 +88,14 @@ android {
 	}
 }
 
-val javadocJar by tasks.creating(Jar::class) {
-	// FIXME: Dokka is failing for this module - temporarily disabled
-	// dependsOn(tasks.getByName("dokkaHtml"))
-	archiveClassifier.set("javadoc")
-	// from("$buildDir/dokka/html")
-}
+enablePublishing {
+	val javadocJar by tasks.creating(Jar::class) {
+		dependsOn(tasks.getByName("dokkaHtml"))
+		archiveClassifier.set("javadoc")
+		from("$buildDir/dokka/html")
+	}
 
-publishing.publications.withType<MavenPublication> {
-	artifact(javadocJar)
+	publications.withType<MavenPublication> {
+		artifact(javadocJar)
+	}
 }
